Syntax updates for models.

+++ Dash/Models/ReportResult.cs
@@ -10,13 +10,13 @@ namespace Dash.Models
         public string DataError { get; set; }
         public string DataSql { get; set; }
         public string Error { get; set; }
-        public int FilteredTotal { get; set; }
+        public long FilteredTotal { get; set; }
         public bool IsOwner { get; set; }
         public int Page { get; set; }
         public int ReportId { get; set; }
         public string ReportName { get; set; }
         public List<object> Rows { get; set; }
-        public int Total { get; set; }
+        public long Total { get; set; }
         public DateTimeOffset UpdatedDate { get; set; }
     }
 }
